requirements.txt
streamlit
pymupdf
numpy
sentence-transformers
groq
python-dotenv